test: split function and sympy tests
